Moved implementations into separate files
- Cleaned up and finished the ParagraphBuilder with Factory in JsiSkParagraphBuilder and ParagraphBuilder.ts
