+++ package/cpp/api/JsiSkApi.h
@@ -24,7 +24,8 @@
 #include "JsiSkMaskFilterFactory.h"
 #include "JsiSkMatrix.h"
 #include "JsiSkPaint.h"
-#include "JsiSkParagraph.h"
+#include "JsiSkParagraphBuilder.h"
+#include "JsiSkParagraphStyle.h"
 #include "JsiSkPath.h"
 #include "JsiSkPathEffect.h"
 #include "JsiSkPathEffectFactory.h"
@@ -44,6 +45,7 @@
 #include "JsiSkShaderFactory.h"
 #include "JsiSkSurfaceFactory.h"
 #include "JsiSkTextBlobFactory.h"
+#include "JsiSkTextStyle.h"
 #include "JsiSkTypeFaceFontProviderFactory.h"
 #include "JsiSkTypeface.h"
 #include "JsiSkTypefaceFactory.h"
